requirements.txt
dify_plugin~=0.0.1b72

psycopg2-binary~=2.9.10
pandas~=2.2.3